Interactive complex heatmap modal dialog

Interactive complex heatmap modal dialog

InteractiveComplexHeatmapModal(
    input, output, session, ht_list, heatmap_id = NULL,
    
    # parameters passed to InteractiveComplexHeatmapOutput()
    title1 = "Original heatmap", title2 = "Selected sub-heatmap",
    width1 = ifelse(layout == "1|(2-3)", 800, 450),
    height1 = ifelse(layout == "1-(2|3)", 700, 350),
    width2 = 370,
    height2 = 350,
    width3 = ifelse(layout == "(1-2)|3", 800, 370),
    layout = ifelse("brush" %in% response, "(1-2)|3", "1-3"), compact = FALSE,
    action = "click", cursor = TRUE, response = c(action, "brush"),
    brush_opt = list(stroke = "#f00", opacity = 0.6),
    output_ui = TRUE, output_ui_float = FALSE,
    
    # parameters passed to makeInteractiveComplexHeatmap()
    click_action = NULL, brush_action = NULL,
    
    # other configurations
    js_code = "", close_button = TRUE, cancel_action = c("remove", "hide"))

Arguments

input
Passed from the Shiny server function.

output
Passed from the Shiny server function.

session
Passed from the Shiny server function.

ht_list
A Heatmap-class or a HeatmapList-class object.

heatmap_id
ID of the plot. If it is not specified, an internal ID is assigned.

title1
Pass to InteractiveComplexHeatmapOutput.

title2
Pass to InteractiveComplexHeatmapOutput.

width1
Pass to InteractiveComplexHeatmapOutput.

height1
Pass to InteractiveComplexHeatmapOutput.

width2
Pass to InteractiveComplexHeatmapOutput.

height2
Pass to InteractiveComplexHeatmapOutput.

width3
Pass to InteractiveComplexHeatmapOutput.

layout
Pass to InteractiveComplexHeatmapOutput.

compact
Pass to InteractiveComplexHeatmapOutput.

action
Pass to InteractiveComplexHeatmapOutput.

cursor
Pass to InteractiveComplexHeatmapOutput.

response
Pass to InteractiveComplexHeatmapOutput.

brush_opt
Pass to InteractiveComplexHeatmapOutput.

output_ui
Pass to InteractiveComplexHeatmapOutput.

output_ui_float
Pass to InteractiveComplexHeatmapOutput.

click_action
Pass to makeInteractiveComplexHeatmap.

brush_action
Pass to makeInteractiveComplexHeatmap.

js_code
Additional JavaScript code that is put after the interactive heatmap UI. The value can be a text or a function that takes "heatmap ID" as the argument and returns the formatted JavaScript code.

close_button
Whether to add a close button at the end of the widget. If it is FALSE, the widget can be closed by clicking outside of the widget.

cancel_action
Whether to remove the UI from HTML or just hide it when the UI is closed.

Details

It creates an interactive heatmap "modal dialog" according to a certain action.

The function is normally put inside observe or observeEvent.

Value

No value is returned.

Examples

if(interactive()) {
    require(ComplexHeatmap)

    ui = fluidPage(
        actionButton("show_heatmap", "Generate_heatmap"),
    )

    server = function(input, output, session) {
        m = matrix(rnorm(100), 10)
        ht = Heatmap(m)

        observeEvent(input$show_heatmap, {
            InteractiveComplexHeatmapModal(input, output, session, ht)
        })
    }
    shiny::shinyApp(ui, server)
}