Examples of the interactive complex heatmaps

Examples of the interactive complex heatmaps

ht_shiny_example(which)

Arguments

which
An integer of which example to use. The list of all examples can be obtained by executing ht_shiny_example with no argument.

Details

The source code of all examples are in systm.file("examples", "examples.R").

Value

No value is returned.

Examples

ht_shiny_example

#> function (which) 
#> {
#>     examples = get_examples()
#>     if (missing(which)) {
#>         cat("There are following examples:\n\n")
#>         title = sapply(examples, function(x) x$title)
#>         for (i in seq_along(title)) {
#>             lines = strwrap(title[i], width = getOption("width") - 
#>                 5)
#>             lines[1] = paste0(ifelse(nchar(i) == 1, " ", ""), 
#>                 i, ". ", lines[1])
#>             lines[-1] = paste0(strrep(" ", nchar(i) + 2), lines[-1])
#>             cat(paste(lines, collapse = "\n"))
#>             cat("\n")
#>         }
#>         cat("\n")
#>     }
#>     else {
#>         which = which[1]
#>         code = examples[[which]]$code
#>         k = which(grepl("rmarkdown::run\\(", code))
#>         if (length(k)) {
#>             eval(parse(text = code[k]))
#>             return(invisible(NULL))
#>         }
#>         library_calls = code[grepl("library\\(.*\\)", code)]
#>         if (length(library_calls)) {
#>             required_pkgs = gsub("library\\((.*)\\)", "\\1", 
#>                 library_calls)
#>             for (pkg in required_pkgs) {
#>                 if (!requireNamespace(pkg, quietly = TRUE)) {
#>                   stop_wrap(paste0("Package", pkg, "should be installed for running this example."))
#>                 }
#>             }
#>         }
#>         cat("Processing the heatmaps... It takes different time depending on examples.\n")
#>         if (any(grepl("ht_shiny\\(", code))) {
#>             code2 = paste(code, collapse = "\n")
#>             code2 = gsub("^\\s+||\\s+$", "", code2)
#>             original_ht_shiny = ht_shiny
#>             ht_shiny = function(ht, ...) {
#>                 html = qq("\n<hr />\n<div>\n<h4>Source code for exporting this Shiny app</h4>\n<pre>\n@{code2}\n</pre>\n</div>")
#>                 original_ht_shiny(ht, ..., html = HTML(html))
#>             }
#>         }
#>         else {
#>             i = which(grepl("shinyApp\\(", code))
#>             code2 = c(code[1:(i - 1)], "", code[seq(i, length(code))])
#>             code_line = paste(code, collapse = "\n")
#>             code_line = gsub("'", "\\\\'", code_line)
#>             code_line = gsub("<", "&lt;", code_line)
#>             code_line = gsub(">", "&gt;", code_line)
#>             code2[i] = qq("\nui = fluidPage(\n    ui,\nHTML('<hr /><div>\n<h4>Source code for exporting this Shiny app</h4>\n<pre>\n@{code_line}\n</pre>\n</div>')\n)\n")
#>             code = code2
#>         }
#>         eval(parse(text = code))
#>     }
#> }
#> <bytecode: 0x7fee3ac9ab10>
#> <environment: namespace:InteractiveComplexHeatmap>

if(interactive()) {
    ht_shiny_example(4)
}