Select an area in the heatmap

Select an area in the heatmap

selectArea(ht_list = get_last_ht(), pos1 = NULL, pos2 = NULL, mark = TRUE, verbose = TRUE,
    ht_pos = NULL, include_annotation = FALSE, calibrate = TRUE)

Arguments

ht_list
A HeatmapList-class object returned by draw,Heatmap-method or draw,HeatmapList-method.

mark
Whether to mark the selected area as a rectangle.

pos1
If the value is NULL, it can be selected by click on the heatmap (of cource, the heatmap should be on the interactive graphic device). If it is set, it must be a unit object with length two which corresponds to the x and y position of the point.

pos2
Another point as pos1, together with pos1 defines the selected region.

verbose
Whether to print messages.

ht_pos
A value returned by htPositionsOnDevice.

include_annotation
Internally used.

calibrate
Internally used.

Details

The regions can be selected interactively or manually by setting pos1 and pos2.

Value

A DataFrame object with row indices and column indices corresponding to the selected region.

Examples

if(dev.interactive()) {
m = matrix(rnorm(100), 10)
rownames(m) = 1:10
colnames(m) = 1:10

ht = Heatmap(m)
ht = draw(ht)
selectArea(ht)

set.seed(123)
ht = Heatmap(m, row_km = 2, column_km = 2)
ht = draw(ht)
selectArea(ht)
}