Process heatmaps on the sever side

Process heatmaps on the sever side

makeInteractiveComplexHeatmap(input, output, session, ht_list,
    heatmap_id = shiny_env$current_heatmap_id,
    click_action = NULL, hover_action = NULL,
    dblclick_action = NULL, brush_action = NULL, res = 72,
    sub_heatmap_cell_fun = NULL, sub_heatmap_layer_fun = NULL)

Arguments

input
Passed from the Shiny server function.

output
Passed from the Shiny server function.

session
Passed from the Shiny server function.

ht_list
A Heatmap-class or a HeatmapList-class object.

heatmap_id
The corresponding heatmap ID from the UI. If there is only one interactive heatmap in the app, this argument does not need to be specified and it will use the current one used in InteractiveComplexHeatmapOutput.

click_action
Additional actions on the server side when receiving a click event on the UI. This self-defined function should accept two or four arguments. If it is two arguments, they should be df and output and if it is four arguments, they should be df, input, output and session.

hover_action
Additional actions at the server side when receiving a hover event on the UI.

dblclick_action
Additional actions at the server side when receiving a dblclick event on the UI.

brush_action
Additional actions at the server side when receiving a brush event on the UI.

res
Resolution of the plot, pass to renderPlot.

sub_heatmap_cell_fun
The cell_fun specifically defined for sub-heatmap.

sub_heatmap_layer_fun
The layer_fun specifically defined for sub-heatmap.

Value

No value is returned.

Examples

if(interactive()) {
    ht = Heatmap(m)
    ht = draw(ht)

    ui = fluidPage(
        InteractiveComplexHeatmapOutput()
    )

    server = function(input, output, session) {
        makeInteractiveComplexHeatmap(input, output, session, ht)
    }

    shiny::shinyApp(ui, server)
}